pyproject.toml
[project]
name = "smartdocs-assistant"
version = "0.1.0"
description = "RAG Chatbot with Conversational Booking"
authors = ["Siddharth Chaudhary <sidartchy203@gmail.com>"]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "langchain>=0.0.350",
    "openai>=1.3.0",
    "faiss-cpu>=1.7.4",
    "pypdf>=3.17.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "google-generativeai>=0.8.0,<0.9.0",
    "langchain-google-genai>=2.0.10",
    "qdrant-client>=1.7.0",
    "redis>=5.0.0",
    "python-dotenv>=1.1.1",
    "phonenumbers>=8.13.0",
    "dateparser>=1.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
